requirements.txt
face_recognition_models
face_recognition
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
matplotlib
